Add wrapper component to record page views
Tracking a page view is a piece of logic separate from what should be
happening in most React components; these components shouldn't need to
mix in the functionality to make that happen.

This new wrapper will automatically record page view events and provides
a mechanism to wait up to a given delay to prevent recording
"accidental" views, such as when someone is simply scanning through the
links.

The heart of this component is separated from the actual analytics
tracking code, making it easy to inject a different recording behavior.