Stop menu name edit button from wrapping
The edit button was wrapping at mobile widths and at 660px to 706px. Mobile is
easy: add a breakpoint. The other one was tougher. The solution was to change
the width of the element to a calc() value that provided space for the button.

This was originally avoided in the pre-oss repo commit 0ef153c because
`text-overflow: ellipsis` did not support `width: calc()`. Browsers appear to
have improved since then.